Reorganize proposal impls modules
