+++ crates/cosmos/cosmos-test-components/src/chain_driver/components.rs
@@ -7,13 +7,13 @@ use hermes_test_components::chain_driver::traits::proposal::poll_status::Proposa
 use hermes_test_components::chain_driver::traits::proposal::query_status::ProposalStatusQuerierComponent;
 
 use crate::chain_driver::impls::amount::GenerateRandomAmount;
-use crate::chain_driver::impls::deposit_proposal::DepositProposalWithChainCommand;
-use crate::chain_driver::impls::poll_proposal_status::PollProposalStatus;
-use crate::chain_driver::impls::proposal_id::ProvideU64ProposalId;
-use crate::chain_driver::impls::proposal_status::{
+use crate::chain_driver::impls::proposal::deposit::DepositProposalWithChainCommand;
+use crate::chain_driver::impls::proposal::poll_status::PollProposalStatus;
+use crate::chain_driver::impls::proposal::query_status::{
     ProvideCosmosProposalStatusType, QueryProposalStatusWithChainCommand,
 };
-use crate::chain_driver::impls::vote_proposal::VoteGovernanceProposalWithChainCommand;
+use crate::chain_driver::impls::proposal::vote::VoteGovernanceProposalWithChainCommand;
+use crate::chain_driver::impls::proposal_id::ProvideU64ProposalId;
 use crate::chain_driver::traits::vote_proposal::GovernanceProposalVoterComponent;
 
 pub struct CosmosChainDriverComponents;
