+++ crates/cosmos/cosmos-test-components/src/chain_driver/impls/mod.rs
@@ -1,6 +1,3 @@
 pub mod amount;
-pub mod deposit_proposal;
-pub mod poll_proposal_status;
+pub mod proposal;
 pub mod proposal_id;
-pub mod proposal_status;
-pub mod vote_proposal;
