+++ crates/cosmos/cosmos-test-components/src/chain_driver/impls/proposal/mod.rs
@@ -1,0 +1,4 @@
+pub mod deposit;
+pub mod poll_status;
+pub mod query_status;
+pub mod vote;
